requirements.txt
fastapi>=0.115.6
uvicorn>=0.34.0
langchain>=0.3.13
langchain_community>=0.3.12
langchain_core>=0.3.27
python_dotenv>=1.0.1
openai>=1.58.1
langchain-openai>=0.2.13